+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed `God Emperor` `Channel Energy` obtained from level 5 -> 3.
 - Changed `God Emperor` `Armor Saint` obtained from level 4 -> 3.
 - Renamed `Energy Immunity Selection` to `Energy Immunity`.
+- Renamed `Aura of Golden Protection` to `Gold Barrier`.
 - Updated description for `Isekai Protagonist` `Fighter Training`.
 - Updated icon for `Exceptional Feats`.
 - Updated icon for `Hero` `Graceful Combat` feature.